extract rich text to separate component
